refactor: adapter 瘦身，open_* Playwright 逻辑回归脚本模块
- adapter: 817→438 行，open_oa_page/open_invoice_page/open_stamp_page
  改为委托 JtnArchiveScript/JtnInvoiceScript/JtnStampScript.open_page()
- archive/playwright_archive: 新增 _open_page + _click_delete_button
- stamp/playwright_stamp: 新增 _open_page
- 新建 invoice/ 模块（playwright_invoice + service）
- mixin 的 _open_page 返回 (playwright, browser)，adapter 管理 _active_browser_sessions
