+++ backend/apps/oa_filing/services/oa_scripts/jtn/adapter.py
@@ -216,414 +216,35 @@ class JTNAdapter(FilingAdapter, StampAdapter, ArchiveAdapter, CaseImportAdapter,
         description: str,
     ) -> None:
         """打开 OA 归档页面，填写案件编号和小结，保持浏览器打开。"""
-        import asyncio
+        from apps.oa_filing.services.oa_scripts.jtn.archive import JtnArchiveScript
 
-        from playwright.async_api import async_playwright
-
-        from apps.oa_filing.services.oa_scripts.jtn.archive.constants import (
-            _POPUP_IFRAME_KEYWORD,
-            AJAX_WAIT,
-            ARCHIVE_PAGE_URL,
-            DESCRIPTION_SELECTOR,
-            IFRAME_SEARCH_FN,
-            MEDIUM_WAIT,
-            POPUP_WAIT,
-            SHORT_WAIT,
-        )
-
-        playwright = await async_playwright().start()
-        browser = await playwright.chromium.launch(headless=False)
-        context = await browser.new_context()
-        page = await context.new_page()
-
-        try:
-            # ── 登录 ──
-            cached = self._auth.load_cookies()
-            if cached:
-                logger.info("使用缓存 cookies 登录 OA")
-                await self._auth.inject_to_context(context, cached)
-                await page.goto(ARCHIVE_PAGE_URL, wait_until="domcontentloaded", timeout=60_000)
-                await asyncio.sleep(MEDIUM_WAIT)
-                if "login" not in page.url.lower():
-                    logger.info("Cookies 有效，已进入归档页面")
-                else:
-                    logger.warning("缓存 cookies 已失效，执行 SSO 扫码登录")
-                    cookies = await self._auth.sso_login()
-                    await self._auth.inject_to_context(context, cookies)
-            else:
-                logger.info("无缓存 cookies，执行 SSO 扫码登录")
-                cookies = await self._auth.sso_login()
-                await self._auth.inject_to_context(context, cookies)
-
-            # ── 导航到归档页面 ──
-            await page.goto(ARCHIVE_PAGE_URL, wait_until="domcontentloaded", timeout=60_000)
-            await asyncio.sleep(MEDIUM_WAIT)
-
-            if "login" in page.url.lower():
-                logger.warning("当前在登录页，等待 SSO 扫码...")
-                await page.wait_for_url("**/ims.jtn.com/projclose/**", timeout=180_000)
-                await asyncio.sleep(MEDIUM_WAIT)
-
-            logger.info("已进入归档页面: %s", page.url)
-
-            # ── 搜索并选择案件 ──
-            if oa_case_number:
-                await self._search_and_select_case(page, oa_case_number)
-
-            # ── 填写案件小结 ──
-            await self._fill_description(page, description)
-
-            # ── 点击删除按钮（清除默认附件） ──
-            await self._click_delete_button(page)
-
-            logger.info("OA 页面已打开并填写完成，浏览器保持打开状态")
-            # 保持 playwright/browser 引用，防止 GC 回收关闭浏览器
-            _active_browser_sessions.append((playwright, browser))
-
-        except Exception:
-            await browser.close()
-            await playwright.stop()
-            raise
+        script = JtnArchiveScript(account=str(credential.account), password=str(credential.password))
+        playwright, browser = await script.open_page(oa_case_number, description)
+        _active_browser_sessions.append((playwright, browser))
 
     async def open_invoice_page(
         self,
         credential: Any,
         oa_case_number: str,
     ) -> None:
-        """打开 OA 发票管理页面，输入案件编号、搜索、点击申请对外开票，保持浏览器打开。"""
-        import asyncio
+        """打开 OA 发票页面，输入案件编号并跳转到开票页面，保持浏览器打开。"""
+        from apps.oa_filing.services.oa_scripts.jtn.invoice import JtnInvoiceScript
 
-        from playwright.async_api import async_playwright
-
-        _INVOICE_URL = (
-            "https://ims.jtn.com/invoice/ProjectFapiaoRequest.aspx"
-            "?FirstModel=FINANCE&SecondModel=FINANCE002&ThirdModel=FINANCE002-04"
-        )
-        _CASE_NO_INPUT = "#ctl00_ctl00_mainContentPlaceHolder_projmainPlaceHolder_project_no"
-        _SEARCH_BTN_XP = '//*[@id="wrap"]/div[1]/div[2]/div/div[4]/div[2]/table/tbody/tr[5]/td[3]/div/a'
-        _FIRST_APPLY_LINK_XP = '//*[@id="wrap"]/div[1]/div[2]/div/div[5]/table/tbody/tr[2]/td[9]/a'
-
-        playwright = await async_playwright().start()
-        browser = await playwright.chromium.launch(headless=False)
-        context = await browser.new_context()
-        page = await context.new_page()
-
-        try:
-            # ── 登录 ──
-            cached = self._auth.load_cookies()
-            if cached:
-                logger.info("使用缓存 cookies 登录 OA")
-                await self._auth.inject_to_context(context, cached)
-                await page.goto(_INVOICE_URL, wait_until="domcontentloaded", timeout=60_000)
-                await asyncio.sleep(2)
-                if "login" not in page.url.lower():
-                    logger.info("Cookies 有效，已进入发票页面")
-                else:
-                    logger.warning("缓存 cookies 已失效，执行 SSO 扫码登录")
-                    cookies = await self._auth.sso_login()
-                    await self._auth.inject_to_context(context, cookies)
-            else:
-                logger.info("无缓存 cookies，执行 SSO 扫码登录")
-                cookies = await self._auth.sso_login()
-                await self._auth.inject_to_context(context, cookies)
-
-            # ── 导航到发票管理页面 ──
-            await page.goto(_INVOICE_URL, wait_until="domcontentloaded", timeout=60_000)
-            await asyncio.sleep(2)
-
-            if "login" in page.url.lower():
-                logger.warning("当前在登录页，等待 SSO 扫码...")
-                await page.wait_for_url("**/ims.jtn.com/invoice/**", timeout=180_000)
-                await asyncio.sleep(2)
-
-            logger.info("已进入发票管理页面: %s", page.url)
-
-            # ── 输入案件编号 ──
-            if oa_case_number:
-                logger.info("输入案件编号: %s", oa_case_number)
-                await page.evaluate(f"""() => {{
-                    const el = document.getElementById("{_CASE_NO_INPUT.lstrip("#")}");
-                    if (el) {{
-                        el.removeAttribute('readonly');
-                        el.removeAttribute('disabled');
-                        el.value = '{oa_case_number}';
-                        el.dispatchEvent(new Event('input', {{ bubbles: true }}));
-                        el.dispatchEvent(new Event('change', {{ bubbles: true }}));
-                    }}
-                }}""")
-                await asyncio.sleep(0.5)
-
-                # ── 点击查找 ──
-                logger.info("点击查找按钮")
-                search_btn = page.locator(f"xpath={_SEARCH_BTN_XP}")
-                count = await search_btn.count()
-                if count > 0:
-                    await search_btn.click(timeout=10_000)
-                else:
-                    logger.warning("XPath 未找到查找按钮，跳过")
-                await asyncio.sleep(3)
-
-                # ── 点击第一个结果的"申请对外开票" ──
-                logger.info("点击申请对外开票")
-                try:
-                    apply_link = page.locator(f"xpath={_FIRST_APPLY_LINK_XP}")
-                    await apply_link.first.click(timeout=10_000)
-                    await asyncio.sleep(2)
-                    logger.info("已跳转到开票页面: %s", page.url)
-                except Exception:
-                    logger.warning("XPath 未找到，尝试 JS 点击")
-                    result = await page.evaluate("""() => {
-                        const links = document.querySelectorAll('#wrap table a');
-                        for (const a of links) {
-                            if (a.textContent.includes('申请') || a.textContent.includes('开票')) {
-                                a.click(); return 'clicked: ' + a.textContent.trim();
-                            }
-                        }
-                        return 'not found';
-                    }""")
-                    logger.info("JS 点击结果: %s", result)
-                    await asyncio.sleep(2)
-
-            logger.info("开票页面已打开，浏览器保持打开状态")
-            # 保持 playwright/browser 引用，防止 GC 回收关闭浏览器
-            _active_browser_sessions.append((playwright, browser))
-
-        except Exception:
-            await browser.close()
-            await playwright.stop()
-            raise
+        script = JtnInvoiceScript(account=str(credential.account), password=str(credential.password))
+        playwright, browser = await script.open_page(oa_case_number)
+        _active_browser_sessions.append((playwright, browser))
 
     async def open_stamp_page(
         self,
         credential: Any,
         oa_case_number: str,
     ) -> None:
-        """打开 OA 盖章页面，登录→搜索案件→填表（所函/公章+电子公章/3份），保持浏览器打开。"""
-        import asyncio
+        """打开 OA 盖章页面，登录→搜索案件→填表，保持浏览器打开。"""
+        from apps.oa_filing.services.oa_scripts.jtn.stamp import JtnStampScript
 
-        from playwright.async_api import async_playwright
-
-        from apps.oa_filing.services.oa_scripts.jtn.stamp.constants import _POPUP_IFRAME_KEYWORD as STAMP_POPUP_KEYWORD
-        from apps.oa_filing.services.oa_scripts.jtn.stamp.constants import AJAX_WAIT as STAMP_AJAX_WAIT
-        from apps.oa_filing.services.oa_scripts.jtn.stamp.constants import (
-            DEFAULT_STAMP_COPIES,
-            FILE_TYPE_SOUHAN,
-            IFRAME_SEARCH_FN,
-        )
-        from apps.oa_filing.services.oa_scripts.jtn.stamp.constants import MEDIUM_WAIT as STAMP_MEDIUM_WAIT
-        from apps.oa_filing.services.oa_scripts.jtn.stamp.constants import POPUP_WAIT as STAMP_POPUP_WAIT
-        from apps.oa_filing.services.oa_scripts.jtn.stamp.constants import SHORT_WAIT as STAMP_SHORT_WAIT
-        from apps.oa_filing.services.oa_scripts.jtn.stamp.constants import (
-            STAMP_PAGE_URL,
-            STAMP_TYPE_INDEX_DIANZI,
-            STAMP_TYPE_INDEX_GONGZHANG,
-            XPATH_FILE_TYPE,
-            XPATH_SEARCH_CASE_BTN,
-            XPATH_STAMP_COPIES,
-        )
-
-        playwright = await async_playwright().start()
-        browser = await playwright.chromium.launch(headless=False)
-        context = await browser.new_context()
-        page = await context.new_page()
-
-        try:
-            # ── 登录 ──
-            cached = self._auth.load_cookies()
-            if cached:
-                logger.info("使用缓存 cookies 登录盖章页面")
-                await self._auth.inject_to_context(context, cached)
-                await page.goto(STAMP_PAGE_URL, wait_until="domcontentloaded", timeout=60_000)
-                await asyncio.sleep(STAMP_MEDIUM_WAIT)
-                if "login" not in page.url.lower():
-                    logger.info("Cookies 有效，已进入盖章页面")
-                else:
-                    logger.warning("缓存 cookies 已失效，执行 SSO 扫码登录")
-                    cookies = await self._auth.sso_login()
-                    await self._auth.inject_to_context(context, cookies)
-            else:
-                logger.info("无缓存 cookies，执行 SSO 扫码登录")
-                cookies = await self._auth.sso_login()
-                await self._auth.inject_to_context(context, cookies)
-
-            # ── 导航到盖章页面 ──
-            await page.goto(STAMP_PAGE_URL, wait_until="domcontentloaded", timeout=60_000)
-            await asyncio.sleep(STAMP_MEDIUM_WAIT)
-
-            if "login" in page.url.lower():
-                logger.warning("当前在登录页，等待 SSO 扫码...")
-                await page.wait_for_url("**/ims.jtn.com/projdoc/**", timeout=180_000)
-                await asyncio.sleep(STAMP_MEDIUM_WAIT)
-
-            logger.info("已进入盖章页面: %s", page.url)
-
-            if not oa_case_number:
-                logger.warning("无案件编号，跳过自动填表")
-                _active_browser_sessions.append((playwright, browser))
-                return
-
-            # ── 搜索并选择案件 ──
-            search_btn = page.locator(XPATH_SEARCH_CASE_BTN)
-            await search_btn.first.click()
-            await asyncio.sleep(STAMP_POPUP_WAIT)
-
-            popup_frame = None
-            for frame in page.frames:
-                if STAMP_POPUP_KEYWORD in frame.url:
-                    popup_frame = frame
-                    break
-            if popup_frame is None:
-                raise RuntimeError("未找到案件搜索弹窗 iframe")
-
-            await popup_frame.evaluate(f"""() => {{
-                const el = document.getElementById("project_no");
-                el.removeAttribute("readonly");
-                el.value = "{oa_case_number}";
-            }}""")
-            await asyncio.sleep(STAMP_SHORT_WAIT)
-            await popup_frame.evaluate(IFRAME_SEARCH_FN)
-            await asyncio.sleep(STAMP_AJAX_WAIT)
-
-            radio_count = await popup_frame.locator('input[type="radio"]').count()
-            if radio_count == 0:
-                raise RuntimeError(f"未找到案件: {oa_case_number}")
-
-            await popup_frame.evaluate('document.querySelectorAll("input[type=radio]")[0].click()')
-            await asyncio.sleep(STAMP_SHORT_WAIT)
-
-            await page.evaluate("""() => {
-                const layers = document.querySelectorAll(".layui-layer");
-                for (const layer of layers) {
-                    for (const a of layer.querySelectorAll("a")) {
-                        if (a.innerText.trim() === "选择") { a.click(); return; }
-                    }
-                }
-            }""")
-            await asyncio.sleep(STAMP_POPUP_WAIT)
-            logger.info("案件已选择: %s", oa_case_number)
-
-            # ── 填写盖章表单 ──
-            file_type = page.locator(XPATH_FILE_TYPE)
-            await file_type.first.select_option(value=FILE_TYPE_SOUHAN)
-            await asyncio.sleep(STAMP_SHORT_WAIT)
-            logger.info("文档类型: 所函")
-
-            for idx in [STAMP_TYPE_INDEX_GONGZHANG, STAMP_TYPE_INDEX_DIANZI]:
-                label = page.locator(f'//*[@id="table_file_1"]/tbody/tr/td[2]/ul/li[{idx}]/label')
-                await label.first.click()
-                await asyncio.sleep(STAMP_SHORT_WAIT)
-            logger.info("盖章类型: 公章, 电子公章")
-
-            copies_input = page.locator(XPATH_STAMP_COPIES)
-            await copies_input.first.fill(str(DEFAULT_STAMP_COPIES))
-            await asyncio.sleep(STAMP_SHORT_WAIT)
-            logger.info("盖章份数: %d", DEFAULT_STAMP_COPIES)
-
-            logger.info("盖章表单已填写完成，浏览器保持打开状态")
-            _active_browser_sessions.append((playwright, browser))
-
-        except Exception:
-            await browser.close()
-            await playwright.stop()
-            raise
-
-    async def _search_and_select_case(self, page: Any, case_no: str) -> None:
-        """搜索案件并选择（复用 archive 的逻辑）。"""
-        import asyncio
-
-        from apps.oa_filing.services.oa_scripts.jtn.archive.constants import (
-            _POPUP_IFRAME_KEYWORD,
-            AJAX_WAIT,
-            IFRAME_SEARCH_FN,
-            POPUP_WAIT,
-            SHORT_WAIT,
-        )
-
-        logger.info("搜索案件: %s", case_no)
-
-        await page.evaluate("""() => {
-            const tds = document.querySelectorAll('td');
-            for (const td of tds) {
-                if (td.textContent.trim().startsWith('案件编号')) {
-                    const row = td.closest('tr');
-                    if (row) { const img = row.querySelector('img'); if (img) { img.click(); return; } }
-                }
-            }
-        }""")
-        await asyncio.sleep(POPUP_WAIT)
-
-        popup_frame = None
-        for frame in page.frames:
-            if _POPUP_IFRAME_KEYWORD in frame.url:
-                popup_frame = frame
-                break
-        if popup_frame is None:
-            raise RuntimeError("未找到案件搜索弹窗 iframe")
-
-        await popup_frame.evaluate(f"""() => {{
-            const el = document.getElementById("project_no");
-            el.removeAttribute("readonly");
-            el.value = "{case_no}";
-        }}""")
-        await asyncio.sleep(SHORT_WAIT)
-
-        await popup_frame.evaluate(IFRAME_SEARCH_FN)
-        await asyncio.sleep(AJAX_WAIT)
-
-        radio_count = await popup_frame.locator('input[type="radio"]').count()
-        if radio_count == 0:
-            raise RuntimeError(f"未找到案件: {case_no}")
-
-        await popup_frame.evaluate('document.querySelectorAll("input[type=radio]")[0].click()')
-        await asyncio.sleep(SHORT_WAIT)
-
-        await page.evaluate("""() => {
-            const layers = document.querySelectorAll(".layui-layer");
-            for (const layer of layers) {
-                for (const a of layer.querySelectorAll("a")) {
-                    if (a.innerText.trim() === "选择") { a.click(); return; }
-                }
-            }
-        }""")
-        await asyncio.sleep(POPUP_WAIT)
-        logger.info("案件已回填到主页面")
-
-    async def _fill_description(self, page: Any, description: str) -> None:
-        """填写案件小结。"""
-        import asyncio
-
-        from apps.oa_filing.services.oa_scripts.jtn.archive.constants import DESCRIPTION_SELECTOR, SHORT_WAIT
-
-        logger.info("填写案件小结: %s", description)
-        result = await page.evaluate(f"""() => {{
-            const ta = document.querySelector('{DESCRIPTION_SELECTOR}');
-            if (!ta) return 'textarea not found';
-            ta.readOnly = false;
-            ta.removeAttribute('readonly');
-            ta.value = '{description}';
-            ta.dispatchEvent(new Event('input', {{ bubbles: true }}));
-            ta.dispatchEvent(new Event('change', {{ bubbles: true }}));
-            return 'ok';
-        }}""")
-        if result != "ok":
-            raise RuntimeError(f"填写案件小结失败: {result}")
-        await asyncio.sleep(SHORT_WAIT)
-
-    async def _click_delete_button(self, page: Any) -> None:
-        """点击删除按钮（清除默认附件行）。"""
-        import asyncio
-
-        from apps.oa_filing.services.oa_scripts.jtn.archive.constants import MEDIUM_WAIT
-
-        logger.info("点击删除按钮")
-        btn = page.locator('//*[@id="tblFiles"]/tbody/tr[3]/td[3]')
-        count = await btn.count()
-        if count == 0:
-            logger.warning("未找到删除按钮")
-            return
-        await btn.first.click()
-        await asyncio.sleep(MEDIUM_WAIT)
-        logger.info("删除按钮已点击")
+        script = JtnStampScript(account=str(credential.account), password=str(credential.password))
+        playwright, browser = await script.open_page(oa_case_number)
+        _active_browser_sessions.append((playwright, browser))
 
     # ==================================================================
     # CaseImportAdapter
