+++ backend/apps/oa_filing/services/oa_scripts/jtn/archive/service.py
@@ -19,3 +19,7 @@ class JtnArchiveScript(PlaywrightArchiveMixin):
     async def run(self, form_data: ArchiveFormData) -> None:
         """执行归档材料提交全流程。"""
         await self._run_archive_submission(form_data)
+
+    async def open_page(self, oa_case_number: str, description: str = "详见卷宗") -> tuple:
+        """打开归档页面并填写，返回 (playwright, browser)。"""
+        return await self._open_page(oa_case_number, description)
