+++ backend/apps/oa_filing/services/oa_scripts/jtn/invoice/__init__.py
@@ -1,0 +1,8 @@
+"""金诚同达 OA 发票申请页面自动化。"""
+
+from __future__ import annotations
+
+from .playwright_invoice import PlaywrightInvoiceMixin
+from .service import JtnInvoiceScript
+
+__all__ = ["JtnInvoiceScript"]
